Updated FileUtil.java
Simply added the methods needed to read in values from as well as write out the data into a simple text file. My next update should include a few methods to do the same for csv file formats. I need to look up what are the common ways of extracting data from a database specifically using MySQL. I may simply write a script and include it into the utility to read in a set of common commands ran in MySQL.